1.02 complete
Added all functionality for assignment 1.02 including the following. Added functionality to generatePaths method so that we look at the gates of surrounding maps to pick our gates so that they match. If one of our maps is on the border of the world, we just write over the gate at the end of that generatingPaths method. Used the probability function given in spec to vary the spawn rates of the buildings. Implemented the I/O and commands n, e, s, w, f x y, q. Should be ready to turn in once I add README and update changelog.