Identified bug in building placement subroutine
One of my while loops to find a place to stick the buildings was infinite looping. I haven't actually fixed it yet, but earlier I deleted the entire file on accident and had to revert back to an old version. I lost less than an hour of work but now I'm paranoid and will be committing more frequently.