Dijkstra's working.
Fixed the bug. Wasn't properly decreasing the key value of the nodes of the heap. Now dijkstras completely fills the map with the shortest paths, I just need to tweak it so we don't look at unwalkable terrain for the given npcs, and make it so when we print the map we don't print the unwalkable terrain.