Almost done with moving to different maps
Currently can go to the northern map (only when using the top left movement), the hard part is done I just need to add the feature for all movement and directions. Also realized the spec wants me to change some of my structs to classes so I still need to do that and implement the fly feature.