Maybe 30% done with 1.08
Have it so you encounter a pokemon with 1/10 chance in tall grass. We pick the pokemon and then assign the level according to manhattan distance. Then we go through and pick all of the possible moves. Then we go through the possible moves and choose the moves that satisfy the chosen level of our pokemon. If there 0 such moves, we increase the level until there aren't 0. Last thing I did was generate IV values.