Have the fight loop for trainers working-ish
The fight loop for trainers works when all you want to do is make moves, but we have some problems when you want to switch pokemon and I haven't implemented the bag yet. Also implemented the function that computes the damage but it isn't giving correct numbers right now (I don't think).